requirements.txt
langchain_anthropic==0.2.4
langchain_community>=0.4.0
langchain_core==0.3.15
langchain_openai==0.2.5
matplotlib==3.9.2
numpy==2.1.3
python-dotenv>=1.0.0
Requests==2.32.3
sentence_transformers==3.1.1
termcolor==2.5.0
torch==2.9.1
tqdm>=4.65.0
transformers==4.44.2